+++ packages/terra-tabs/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 * Added `responsiveTo` prop to provide ability to consumers to decide whether or not tabs need to completely
 collapse into a menu on smaller viewports.
+* Added orion-fusion-theme and generated the respective reference wdio screenshots.
 
 ### Changed
 * Center-align text label in the single dropdown tab when all tabs completely collapse into a menu.